V15.02.A-1
Fixes to bring Watchdog CEWS and Watchdog CEWS(NC) in line with Stock BTA...

3 barreled AC/2 primarily for T-Bone's Corsair (which will come in a future update

The Stag returns in 4 versions to my mods... the default Clan Wolverine STAG STG-14G, STG-24G Stag II and the -14 and -24GJ which are my modernized versions.   Still have to fix the Preta-Stag issue (also comming in a future update.     Clan Thermal Exchangers base and Mk1 included.

Most importantly with this upload    The Stormback Hunchback HBK-IIC-86 arrives in public release...  With a slight nerf to the clan Stormbreaker the HBK-II-86 is now no longer super duper OP.