V19 Update
This is the V19 Update   It was created by reverse engineering V19 and theirfor may have **MNIOR** issues.  Please DM me directly on Discord for issues rather than clogging the mod channel with them.

As always Delete all Pafftek and NPJointPRoduction folders and install clean as there are always minor structural changes to this mod!