Bug Fix and Patch Update
This is a bugfix that does the following:

FINALLY Fixes the Stag JSON Merges that were non functional in prior releases (which could result in 2 right arm lower actuators!)

Fixes Black Carapace to not work with most Stealth (or Stelf if you prefer) stuff (it is already stealth it doesn't need to be more stealth.

Fixes many descriptions that were incomplete, not visible etc.

Error fixes the Maelstrom, the Omegas and the Mauler MAL-1RJ